Cleanup read_experiment function
Removed a bunch of conditionals. I now use the nancheck lambda
that sets the key to None if there the yaml file entry is null,
or updates the dictionary if the value is non-null. This was done
to make the code easier to read, so it doesn't add any
new (intentional) functionality.