Add descriptive TODO
Multiprocessing in python is weirdly complicated.
The shareddict object is a bad way to track state across
multiple processes. thankfully, the manager.dict object
does the right thing when the keys-values being updated are
independent of each other. In the counter case, it seems like
not all processes have access to the same counter object.
This is very non-intuitive when it seems like the user has
explicitly created an object to share state.